game/requirements.txt
pygame==2.3.0
numpy==1.24.2
moderngl==5.8.1
PyGLM==2.7.0
freetype-py==2.3.0